추천기능

+++ common/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from django.contrib.auth import authenticate, login, update_session_auth_hash
+from django.contrib.auth import authenticate, login, update_session_auth_hash, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import PasswordChangeForm
 from django.contrib import messages
@@ -71,10 +71,13 @@ def password_change(request):
 
 @login_required(login_url='common:login')
 def profile_delete(request):
-    user = get_object_or_404(User, username=request.user.username)
     if request.POST.get('delete') == '회원탈퇴':
-        user.delete()
-        return redirect('footstep:index')
-    else:
-        messages.error(request, '탈퇴하는데 실패했습니다')
-        return redirect('common:profile')
+        username = request.POST['username']
+        password = request.POST['password']
+        user = authenticate(username=username, password=password)
+        if user:
+            user.delete()
+            logout(request)
+            return redirect('footstep:index')
+    messages.error(request, '요청을 실패했습니다')
+    return render(request, 'common/profile.html')
